Option to return bootstrap replica for `mzR` cnstructors
The constructor functions for mzR objects (Mean, Median, Total,
GroupSize, GroupRate) now all provide a Parameter `replicates`,
which will include a vector of bootstrap replicates to the
`mzR` object.

There is also a new function called `getReplicates`, which returns
the replicates as a table. Usage examples can be found in
`?getReplicates`.